// src/app/earlygame/models/enums.tsx
export enum OreTypes {
    Stone = "Stone Ore",
    Copper = "Copper Ore",
    Bronze = "Bronze Ore",
    Silver = "Silver Ore",
    Gold = "Gold Ore",
    Platinum = "Platinum Ore",
    Diamond = "Diamond Ore",
}

export enum GemTypes {
    Ruby,
    Sapphire,
    Emerald,
    Diamond,
    Amethyst,
}

export enum PickaxeTypes {
    Wood = "Wood Pickaxe",
    Stone = "Stone Pickaxe",
    Copper = "Copper Pickaxe",
    Bronze = "Bronze Pickaxe",
    Silver = "Silver Pickaxe",
    Gold = "Gold Pickaxe",
    Platinum = "Platinum Pickaxe",
    Diamond = "Diamond Pickaxe",
}